feat: shared/ui 생성
