+++ src/shared/ui/Input/index.tsx
@@ -1,0 +1,15 @@
+import * as React from "react"
+import { forwardRef } from "react"
+
+type InputProps = React.InputHTMLAttributes<HTMLInputElement> & { className?: string }
+export const Input = forwardRef<HTMLInputElement, InputProps>(({ className = "", type, ...props }, ref) => {
+  return (
+    <input
+      className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+      ref={ref}
+      type={type}
+      {...props}
+    />
+  )
+})
+Input.displayName = "Input"
